+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## UPCOMING
 **_Add new changes here as they land_**
 
+- Selector cache lookup optimization (#1720)
+
 ## 0.7.1 (2022-04-12)
 
 ### Typing